packages/backend/pyproject.toml
[project]
name = "riffroom-backend"
version = "0.1.0"
description = "ML-powered stem separation backend for RiffRoom"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.12",
    "python-decouple>=3.8",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.29.0",
    "demucs>=4.0.1",
    "librosa>=0.10.2",
    "scipy>=1.14.1",
    "numpy>=2.0.2",
    "torch>=2.2.0",
    "websockets>=13.0",
    "aiosqlite>=0.20.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "pyright>=1.1.380",
]
performance = [
    "onnxruntime>=1.18.0",  # Performance optimization - install when cp314 wheels available
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "PT"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
